Working version able to be submitted just update readme. I think we should hanimate the cows if possible

+++ hw2/fence.h
@@ -2,7 +2,7 @@
 #define FENCE_H
 
 //prototypes
-void createFence(double x_posts, double y_posts, double max_x, double min_x, double max_y, double min_y);
+void createFence(int x_posts, int y_posts, double max_x, double min_x, double max_y, double min_y);
 void drawBeam();
 void drawPost();
